Use PyYAML parsing for command-line arguments instead of Pydantic
The use of Pydantic to validate command-line arguments based on the types PyYAML parses from the config file might result in confusing behavior, particularly if the type of the (ultimately unused) entry in the config file is incorrect. Separating out the parsing of the config file and command-line arguments and using PyYAML for both is simpler and probably less confusing. In the future it might make sense to use Pydantic with actual data models to validate both the config file and command-line arguments.